working logout button in profile side
