+++ RoomMates/Views/Expenses/ExpenseView.swift
@@ -1,0 +1,15 @@
+//
+//  ExpenseView.swift
+//  RoomMates
+//
+//  Created by callum on 2024-10-05.
+//
+import SwiftUI
+
+struct ExpensesView: View {
+    var body: some View {
+        Text("Expenses View")
+            .font(.largeTitle)
+            .padding()
+    }
+}
