+++ RoomMates/Views/Feedback/FeedbackView.swift
@@ -1,0 +1,19 @@
+//
+//  FeedbackView.swift
+//  RoomMates
+//
+//  Created by callum on 2024-10-05.
+//
+import SwiftUI
+
+struct FeedbackView: View {
+    var body: some View {
+        Text("Feedback View")
+            .font(.largeTitle)
+            .padding()
+    }
+}
+
+#Preview {
+    FeedbackView()
+}
