+++ RoomMates/Views/HouseInformation/HouseInfoView.swift
@@ -1,0 +1,15 @@
+//
+//  HouseInfoView.swift
+//  RoomMates
+//
+//  Created by callum on 2024-10-05.
+//
+import SwiftUI
+
+struct HouseInfoView: View {
+    var body: some View {
+        Text("House Info View")
+            .font(.largeTitle)
+            .padding()
+    }
+}
